+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 * Fixed an issue where the chosen output format would not be used when exporting.
 
+### Changed
+* The footer now displays details of the error if a file fails to convert properly.
+
 ### 0.2.1 - 2023-08-25
 ### Fixed
 * Fixed an issue that prevented exporting with mipmaps forced to "Disabled".